display for ld better and test_champs folder

+++ srcs/arena/process.c
@@ -80,7 +80,7 @@ void	execute_processes(t_arena *arena)
 			if (opcode_to_mask(it->current_op->opcode) & TEST_PRINT)
 			{
 				ft_printf("\n\n");
-				ft_printf("cycle : \t%lu\n", arena->cycle);
+				ft_printf("cycle : \t%d\n", (int)arena->cycle);
 				ft_printf("PC:\t\t%d\n", it->pc);
 			}
 			execute_process(arena, it);
